refactor(shared): extract request + ApiError
Both apps had byte-identical fetch wrappers and ApiError classes. Move
to @qzr/shared as createApiClient(baseUrl) factory + ApiError class.
Each app constructs one with its own baseUrl, re-exports ApiError so
existing consumers of '../api' keep working.

Adds DOM lib to shared tsconfig so fetch/RequestInit resolve.
