+++ apps/web/src/api.ts
@@ -1,34 +1,10 @@
-import { MeetRole } from '@qzr/shared'
+import { ApiError, createApiClient, MeetRole } from '@qzr/shared'
 
 declare const __API_URL__: string
 
-const baseUrl = __API_URL__ || ''
+const request = createApiClient(__API_URL__ || '')
 
-async function request<T>(path: string, init?: RequestInit): Promise<T> {
-  const res = await fetch(`${baseUrl}${path}`, {
-    ...init,
-    credentials: 'include',
-    headers: {
-      'Content-Type': 'application/json',
-      ...init?.headers,
-    },
-  })
-  if (!res.ok) {
-    const body = await res.json().catch(() => ({}))
-    throw new ApiError(res.status, (body as { error?: string }).error ?? res.statusText)
-  }
-  return res.json() as Promise<T>
-}
-
-export class ApiError extends Error {
-  constructor(
-    public status: number,
-    message: string,
-  ) {
-    super(message)
-    this.name = 'ApiError'
-  }
-}
+export { ApiError }
 
 // ---- Types ----
