+++ packages/shared/src/apiClient.ts
@@ -1,0 +1,24 @@
+export class ApiError extends Error {
+  constructor(
+    public status: number,
+    message: string,
+  ) {
+    super(message)
+    this.name = 'ApiError'
+  }
+}
+
+export function createApiClient(baseUrl: string) {
+  return async function request<T>(path: string, init?: RequestInit): Promise<T> {
+    const res = await fetch(`${baseUrl}${path}`, {
+      ...init,
+      credentials: 'include',
+      headers: { 'Content-Type': 'application/json', ...init?.headers },
+    })
+    if (!res.ok) {
+      const body = await res.json().catch(() => ({}))
+      throw new ApiError(res.status, (body as { error?: string }).error ?? res.statusText)
+    }
+    return res.json() as Promise<T>
+  }
+}
