+++ packages/shared/tsconfig.json
@@ -6,7 +6,7 @@
     "target": "ESNext",
     "module": "ESNext",
     "moduleResolution": "Bundler",
-    "lib": ["ESNext"]
+    "lib": ["ESNext", "DOM"]
   }
 }
